requirements.txt
matplotlib~=3.7.3
numpy~=1.25.2
opencv-python~=4.8.0.76
Pillow~=10.0.0
requests~=2.31.0
torch~=2.0.1
torchvision~=0.15.2
kornia~=0.7.0
rasterio~=1.3.8
tqdm~=4.66.1
scipy~=1.11.2
patchify~=0.2.3
plotly~=5.17.0
scikit-learn~=1.3.1
ttach~=0.0.3
kaleido~=0.2.1